refactor: extract publish to its own module
